♻️ Fold ACPI PM into acpi_wrapper.c, dedupe table walk
The PM helpers were sitting in a separate acpi_pm.c with their own
init wrapper, and acpi_early_init re-implemented the XSDT/RSDT walk
that cosmos_acpi_scan_table already does. Collapse both:

- Move cosmos_acpi_shutdown / cosmos_acpi_reset into acpi_wrapper.c
  alongside the rest of the LAI integration.
- acpi_early_init now finds MADT/MCFG via cosmos_acpi_scan_table.
- Drop the cosmos_acpi_init_namespace wrapper; inline the one-shot
  lai_create_namespace() into the shutdown path where it's used.
- Delete the now-empty acpi_pm.c.
