+++ src/Cosmos.Kernel.Native.MultiArch/ACPI/acpi_wrapper.c
@@ -7,6 +7,7 @@
 #include <stddef.h>
 
 #include <lai/core.h>
+#include <lai/helpers/pm.h>
 #include <acpispec/tables.h>
 
 #define NULL_PTR ((void*)0)
@@ -20,6 +21,7 @@ extern void __cosmos_serial_write_dec_u64(uint64_t value);
 
 // Cosmos support
 extern void cosmos_acpi_set_rsdp(void* rsdp);
+extern void* cosmos_acpi_get_rsdp(void);
 
 // ============================================================================
 // ARM64 GIC structures (MADT subtable types 0x0B-0x0F)
@@ -134,7 +136,6 @@ static inline int sig_eq(const char* a, const char* b) {
 void* cosmos_acpi_scan_table(const char* signature, size_t index) {
     if (signature == NULL_PTR) return NULL_PTR;
 
-    extern void* cosmos_acpi_get_rsdp(void);
     acpi_rsdp_t* rsdp = (acpi_rsdp_t*)cosmos_acpi_get_rsdp();
     if (rsdp == NULL_PTR) return NULL_PTR;
 
@@ -396,50 +397,10 @@ void acpi_early_init(void* rsdp_address, uint64_t hhdm_offset) {
     lai_set_acpi_revision(acpi_rev);
     cosmos_acpi_set_rsdp(rsdp_address);
 
-    // Find ACPI tables (MADT, MCFG) in XSDT or RSDT
-    acpi_header_t* madt = NULL_PTR;
-    acpi_header_t* mcfg = NULL_PTR;
-
-    if (rsdp->revision >= 2 && ((acpi_xsdp_t*)rsdp)->xsdt != 0) {
-        uint64_t xsdt_phys = ((acpi_xsdp_t*)rsdp)->xsdt;
-        acpi_xsdt_t* xsdt = (acpi_xsdt_t*)phys_to_virt(xsdt_phys);
-        __cosmos_serial_write("[ACPI] XSDT at phys 0x");
-        __cosmos_serial_write_hex_u64(xsdt_phys);
-        __cosmos_serial_write(" -> virt 0x");
-        __cosmos_serial_write_hex_u64((uint64_t)xsdt);
-        __cosmos_serial_write("\n");
-
-        uint32_t count = (xsdt->header.length - sizeof(acpi_header_t)) / sizeof(uint64_t);
-        for (uint32_t i = 0; i < count; i++) {
-            acpi_header_t* tbl = (acpi_header_t*)phys_to_virt(xsdt->tables[i]);
-            char* sig = tbl->signature;
-            if (sig[0] == 'A' && sig[1] == 'P' && sig[2] == 'I' && sig[3] == 'C') {
-                madt = tbl;
-                __cosmos_serial_write("[ACPI] MADT found\n");
-            } else if (sig[0] == 'M' && sig[1] == 'C' && sig[2] == 'F' && sig[3] == 'G') {
-                mcfg = tbl;
-                __cosmos_serial_write("[ACPI] MCFG found\n");
-            }
-        }
-    } else if (rsdp->rsdt != 0) {
-        acpi_rsdt_t* rsdt = (acpi_rsdt_t*)phys_to_virt((uint64_t)rsdp->rsdt);
-        __cosmos_serial_write("[ACPI] RSDT at: 0x");
-        __cosmos_serial_write_hex_u32(rsdp->rsdt);
-        __cosmos_serial_write("\n");
-
-        uint32_t count = (rsdt->header.length - sizeof(acpi_header_t)) / sizeof(uint32_t);
-        for (uint32_t i = 0; i < count; i++) {
-            acpi_header_t* tbl = (acpi_header_t*)phys_to_virt((uint64_t)rsdt->tables[i]);
-            char* sig = tbl->signature;
-            if (sig[0] == 'A' && sig[1] == 'P' && sig[2] == 'I' && sig[3] == 'C') {
-                madt = tbl;
-                __cosmos_serial_write("[ACPI] MADT found\n");
-            } else if (sig[0] == 'M' && sig[1] == 'C' && sig[2] == 'F' && sig[3] == 'G') {
-                mcfg = tbl;
-                __cosmos_serial_write("[ACPI] MCFG found\n");
-            }
-        }
-    }
+    acpi_header_t* madt = (acpi_header_t*)cosmos_acpi_scan_table("APIC", 0);
+    acpi_header_t* mcfg = (acpi_header_t*)cosmos_acpi_scan_table("MCFG", 0);
+    if (madt) __cosmos_serial_write("[ACPI] MADT found\n");
+    if (mcfg) __cosmos_serial_write("[ACPI] MCFG found\n");
 
     if (madt) {
         __cosmos_serial_write("[ACPI] Parsing MADT...\n");
@@ -476,3 +437,38 @@ const acpi_gic_info_t* acpi_get_gic_info(void) {
 const acpi_mcfg_info_t* acpi_get_mcfg_info(void) {
     return g_initialized ? &g_mcfg_info : NULL_PTR;
 }
+
+// ============================================================================
+// Power management — ACPI _S5 / FADT reset via LAI (x86 only)
+// ============================================================================
+
+#ifdef ARCH_X64
+
+// Lazily build the AML namespace on first PM call. lai_acpi_reset() reads the
+// FADT directly via laihost_scan and doesn't need the namespace, but
+// lai_enter_sleep(5) must resolve \\_S5_ from the DSDT.
+static int g_namespace_created = 0;
+
+int cosmos_acpi_shutdown(void) {
+    if (cosmos_acpi_get_rsdp() == NULL_PTR) return 1;
+    if (!g_namespace_created) {
+        __cosmos_serial_write("[ACPI-PM] Creating AML namespace...\n");
+        lai_create_namespace();
+        g_namespace_created = 1;
+    }
+    __cosmos_serial_write("[ACPI-PM] lai_enter_sleep(S5)\n");
+    return (int)lai_enter_sleep(5);
+}
+
+int cosmos_acpi_reset(void) {
+    __cosmos_serial_write("[ACPI-PM] lai_acpi_reset\n");
+    return (int)lai_acpi_reset();
+}
+
+#else
+
+// LAI's PM helpers are x86-only (port I/O via FADT PM1a). ARM64 uses PSCI.
+int cosmos_acpi_shutdown(void) { return 1; }
+int cosmos_acpi_reset(void) { return 1; }
+
+#endif
